requirements.txt
streamlit
langchain-core
langchain-community
python-dotenv
gtts
pygame
